requirements.txt
tensorflow-gpu==2.4.0
keras==2.4.3
